pyproject.toml
[tool.poetry]
name = "myHPI"
version = "0.0.0" # automatically set by poetry-dynamic-versioning
description = ""
authors = ["FSR DE <fachschaftsrat@hpi.de>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
wagtail = "^6.2"
django-environ = "^0.11.2"
wagtail-localize = "^1.9"
mozilla-django-oidc = "^4.0.0"
django-bootstrap-icons = "^0.9.0"
django-select2 = "^8.2.0"
django-static-precompiler = {extras = ["libsass"], version = "^2.4"}
django-debug-toolbar = "^4.4.6"
django-permissionedforms = "^0.1"
html2text = "^2020.1.16"
wagtail-markdown = "^0.12.1"
autoflake = "^2.2.1"
setuptools-scm = "^8.0.4"
django-prometheus = "^2.3.1"
djlint = "^1.35.2"
mailmanclient = "^3.3.5"

[tool.poetry.group.dev.dependencies]
pylint = "^3.2.6"
black = "^24.8"
isort = "^5.12.0"
autoflake = "^2.2.1"
pre-commit = "^3.8.0"
pytest = "^8.2.2"
coverage = "^7.6.0"

[tool.poetry.extras]
pgsql = ["psycopg2"]
mysql = ["mysqlclient"]

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.black]
line-length = 100
# https://github.com/psf/black/blob/master/docs/compatible_configs.md

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100
skip_gitignore = true

[tool.pylint.messages_control]
# C0330 and C0301 are disabled for use of black
disable = """C0330, C0301,
duplicate-code, attribute-defined-outside-init, missing-module-docstring, missing-class-docstring,
no-member, invalid-name, import-outside-toplevel, unused-argument, too-many-ancestors, missing-function-docstring,
too-few-public-methods, no-self-use, too-many-arguments, cyclic-import, inconsistent-return-statements,
useless-object-inheritance, logging-fstring-interpolation
"""

[tool.pylint.format]
max-line-length = "100"

[tool.poetry-dynamic-versioning]
enable = true

[tool.djlint]
ignore = "T002,T003,H006,H023,H031,H037"
preserve_blank_lines = true
custom_blocks = "blocktrans,cache"
max_line_length=80
